backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[project]
name = "grove"
version = "2.0.0"
description = "Grove by SEOasis — client onboarding and content fulfillment platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "python-json-logger>=2.0.7",
    "httpx>=0.26.0",
    "redis>=5.0.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.0",
    # Document parsing
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    # Email
    "aiosmtplib>=3.0.0",
    # WebSocket client
    "websockets>=12.0",
    # Form data / file uploads
    "python-multipart>=0.0.6",
    # Sync DB driver for APScheduler job store
    "psycopg2-binary>=2.9.0",
    "greenlet>=3.3.1",
    # S3 storage
    "boto3>=1.34.0",
    # HTML parsing for content extraction
    "beautifulsoup4>=4.12.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.26.0",
    "factory-boy>=3.3.0",
    "faker>=22.0.0",
    "aiosqlite>=0.19.0",
    "ruff>=0.14.0",
    "mypy>=1.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["app"]
omit = ["app/alembic/*", "app/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

# ---------------------------------------------------------------------------
# Ruff Configuration
# ---------------------------------------------------------------------------

[tool.ruff]
target-version = "py312"
line-length = 88
src = ["app", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults (FastAPI Depends)
    "B904",   # raise from within except (too noisy)
    "ARG001", # unused function argument (common in FastAPI)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG001", "ARG002"]
"alembic/*" = ["ARG001"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

# ---------------------------------------------------------------------------
# MyPy Configuration
# ---------------------------------------------------------------------------

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
show_column_numbers = true
pretty = true
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

# Per-module overrides for third-party libraries without stubs
[[tool.mypy.overrides]]
module = [
    "alembic.*",
    "asyncpg.*",
    "redis.*",
    "uvicorn.*",
    "pydantic_settings",
    "pydantic_settings.*",
    "pythonjsonlogger",
    "pythonjsonlogger.*",
    "httpx.*",
    "apscheduler.*",
    "pypdf",
    "pypdf.*",
    "docx",
    "docx.*",
    "aiosmtplib",
    "aiosmtplib.*",
    "websockets",
    "websockets.*",
    "boto3",
    "boto3.*",
    "botocore",
    "botocore.*",
    "bs4",
    "bs4.*",
]
ignore_missing_imports = true

# Allow subclassing Any for modules that use third-party base classes
[[tool.mypy.overrides]]
module = [
    "app.core.config",
    "app.core.logging",
]
disallow_subclassing_any = false

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
